game_type added bwapi.ini file - works for single player and udp lan games.
git-svn-id: https://bwapi.googlecode.com/svn/trunk@2287 273ecf2c-f24b-0410-b6a7-e9298bb4873f

+++ bwapi/BWAPI/Source/BW/Dialog.cpp
@@ -128,8 +128,8 @@ namespace BW
       {
         if ( this->u.list.pbStrFlags[i] == 0 && this->u.list.pdwData[i] == dwValue)
         {
-          this->u.list.bCurrStr       = i;
-          this->u.list.bSelectedIndex = i;
+          this->u.list.bCurrStr       = (BYTE)i;
+          this->u.list.bSelectedIndex = (BYTE)i;
         }
       }
     } // check
